third_party/WebKit/LayoutTests/shadow-dom/v0/closed-mode-deep-combinators-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
PASS result.length is 7
PASS node.id is "openhost"
PASS node.id is "closedhost"
PASS node.id is "v0host"
PASS node.id is "v0host_in_v0host"
PASS node.id is "div_v0_in_v0"
PASS node.id is "openhost_in_v0host"
PASS node.id is "closedhost_in_v0host"
PASS successfullyParsed is true

TEST COMPLETE
